First commit
the UserInfo model is the one that was asked for in the project, I found some people doing similar projects and followed along
I still need to figure out the best way to do TN length validation, if it should be a string because of the special characters, 
or if I can use Regex so they never get passed in and force the length to be 10 with built in min and max.
I have Mocha installed on another project, but am trying not to go too many directions at once.  Stack currently is MongoDB,
with Mongoose, express, and node.js.  I am using nodemon locally, but don't have it listed as a dependency.